fixed sectionsModel comments
